+++ Bank ATM/BankTest/UnitTest1.cs
@@ -10,12 +10,12 @@ namespace BankTest
         [Theory]
         [InlineData(1, 9999)]
         [InlineData(5, 9995)]
-        public void CanWithdraw(double value, double expectedResult)
+        public void CanWithdraw(decimal value, decimal expectedResult)
         {
             //Act-resetting balance to 10000 because of global balance
             Balance = 10000;
             //Arrange
-            double remainingBalance = WithdrawalFromBalance(value);
+            decimal remainingBalance = WithdrawalFromBalance(value);
             //Assert
             Assert.Equal(remainingBalance, expectedResult);
         }
